third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/embed-tag-code-attribute-2-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cembed%20code=//localhost:8000/fictional.swf%20allowscriptaccess=always%3E%3C/embed%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
